+++ app/controllers/api/v2/admin/partners_controller.rb
@@ -1,8 +1,8 @@
 class Api::V2::Admin::PartnersController < Api::V2::Admin::ContactsController
 
   def index
-    partners = ::Partner.all
-    if partners&.empty? or partners.nil? and ::Partner.all.size > 0
+    partners = policy_scope [:admin, ::Partner]
+    if partners&.empty? or partners.nil?
       render json: partners, status: 401
     else
       partners = if params[:name] && params[:federal_registration] 
@@ -16,16 +16,16 @@ class Api::V2::Admin::PartnersController < Api::V2::Admin::ContactsController
 
   def show
     partner = ::Partner.find(params[:id])
-    # authorize partner
+    authorize [:admin, partner]
     render json: partner, status: 200
   end
 
   def create
     partner = ::Partner.new(partner_params)
-    # authorize partner
+    authorize [:admin, partner]
     if partner.save
       update_contact(partner)
-      update_user(partner)
+      update_user(partner, fdr_old, fdr_new)
       premio_ideal(partner)
       render json: partner, status: 201
     else
@@ -35,7 +35,7 @@ class Api::V2::Admin::PartnersController < Api::V2::Admin::ContactsController
 
   def update
     partner = ::Partner.find(params[:id])
-    # authorize partner
+    authorize [:admin, partner]
     fdr_old = partner.federal_registration
     fdr_new = partner_params['federal_registration']
     if partner.update(partner_params)
@@ -50,22 +50,17 @@ class Api::V2::Admin::PartnersController < Api::V2::Admin::ContactsController
 
   def destroy
     partner = ::Partner.find(params[:id])
-    # authorize partner
+    authorize [:admin, partner]
     user = partner.user
     partner.destroy
-    begin
-      user.destroy unless partner.user.client && partner.user.company && partner.user.employee
-    rescue
-    end
+    user.destroy if !partner.user.client
     head 204
   end
 
   private
 
   def partner_params
-      params.permit(:id, :name, :federal_registration, :state_registration, 
-      :kind, :active, :started_date, :renewal_date, :description, :origin, :percent, :agency, 
-      :ocupation, :account, :favored, :user_id, :bank_id, :discount3, :discount5, :discount8, :cash_redemption)
+    params.permit(policy([:admin, ::Partner]).permitted_attributes)
   end
 
   def update_user(partner, fdr_old = 'new', fdr_new = 'new')
